refactor(cli): extract the dev process supervisor to its own module
runDevSupervisor — the fastagent dev watch-and-restart mechanism (~75 lines: spawn the CLI as a
worker, IPC ready signal, chokidar watch, debounced reload, SIGINT/SIGTERM shutdown) — lived inline in
the 690-line cli.ts dispatch file. It is process orchestration, a distinct concern from command
dispatch, and it only needs the workspace dir. Moved verbatim to src/dev-supervisor.ts as
runDevSupervisor(dir); cli.ts imports it and drops the now-unused chokidar import. No behavior change
(cli.ts 690 -> 606); dev reload + Ctrl-C worker cleanup verified by smoke.
